Simplify zookeeper selector
As there can only be one wondering trader loaded at any one time it's
unlikely for there to be a conflict in selecting them. Removing the name
simplyfies this and also fixes an issue (it seems) were advancements
didn't always work at night.

In addition we now use the mob id as the icon player head owner id to
clean up the gif diff from now on.